Remove util function to open image in lightbox
