--- src/modules/util/files/index.ts
@@ -1,1 +1,0 @@
-export { openImageInLightbox } from "./image-lightbox";
